chapter_001/src/test/java/ua/learningjava/condition/DummyBotTest.java
package ua.learningjava.condition;

import org.junit.Test;
import static org.hamcrest.core.Is.is;
import static org.junit.Assert.assertThat;

/**
 * @author Avaltrois
 * @version $Id$
 * @since 0.1
 */
public class DummyBotTest {
    @Test
    public void whenGreetBot() {
        DummyBot bot = new DummyBot();
        assertThat(
                bot.answer("Hey Bot"),
                is("Hello Man")
        );
    }

    @Test
    public void whenByuBot() {
        DummyBot bot = new DummyBot();
        assertThat(
                bot.answer("Bye Bot"),
                is("Bye Man")
        );
    }

    @Test
    public void whenUnknownBot() {
        DummyBot bot = new DummyBot();
        assertThat(
                bot.answer("How much is 2-2?"),
                is("It baffles me. Ask another question.")
        );
    }
}
